requirements.txt
# Core LangChain and LangGraph dependencies
langchain>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.20
langchain-core>=0.1.0

# FastAPI and server dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# OpenAI API client
openai>=1.0.0

# HTTP client for examples
requests>=2.31.0

# Environment variable management
python-dotenv>=1.0.0

# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Standard library dependencies (usually included, but listed for clarity)
# typing-extensions (for Annotated)
typing-extensions>=4.8.0

